Restructured file structure to fit Ruben's configurations, deleted some redundant services
