+++ couch-critics-frontendBETA/cc/src/app/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthenticationService} from 'src/app/authentication.service'
+import { AuthenticationService } from '../services/authentication.service';
 
 @Component({
   selector: 'app-login',
