+++ couch-critics-frontendBETA/cc/src/app/services/authentication.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { User } from '../app/models/user';
-import { environment as env } from '../environments/environment';
+import { User } from '../models/user';
 import { map } from 'rxjs/operators';
+import { environment as env } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -11,6 +11,10 @@ import { map } from 'rxjs/operators';
 export class AuthenticationService {
 
   constructor(private http: HttpClient) { }
+
+  //Change to env.H2_TEST to access H2-CONSOLE URL
+  //Change to env.BACKEND_URL if you need to access backend
+  private apiServerUrl = env.BACKEND_URL;
 
 
   login(username:string, password:string): Observable<User>{
@@ -20,9 +24,7 @@ export class AuthenticationService {
     }
 
 
-    //We are currently testing against the H2-CONSOLE URL
-    //Change to BACKEND_URL if you need to access backend
-    return this.http.post(`${env.TEST_H2_URL}/auth`, loginInfo).pipe(
+    return this.http.post(`${this.apiServerUrl}/auth`, loginInfo).pipe(
       map(response => response as User)
     );
   }
