+++ couch-critics-frontendBETA/cc/src/app/services/manager-service.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ManagerServiceService {
+
+  constructor() { }
+}
